requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
transformers>=4.40.0
torch>=2.2.0
pdfplumber>=0.11.0
pymupdf>=1.24.0
python-docx>=1.1.0
